move traits outside of feature flags.
