+++ canhttp/src/http/request.rs
@@ -4,28 +4,10 @@ use ic_cdk::api::management_canister::http_request::{
     HttpMethod as IcHttpMethod, TransformContext,
 };
 use thiserror::Error;
+use crate::{MaxResponseBytesRequestExtension, TransformContextRequestExtension};
 
 /// HTTP request with a body made of bytes.
 pub type HttpRequest = http::Request<Vec<u8>>;
-
-/// Add support for max response bytes.
-pub trait MaxResponseBytesRequestExtension: Sized {
-    /// Set the max response bytes.
-    ///
-    /// If provided, the value must not exceed 2MB (2_000_000B).
-    /// The call will be charged based on this parameter.
-    /// If not provided, the maximum of 2MB will be used.
-    fn set_max_response_bytes(&mut self, value: u64);
-
-    /// Retrieves the current max response bytes value, if any.
-    fn get_max_response_bytes(&self) -> Option<u64>;
-
-    /// Convenience method to use the builder pattern.
-    fn max_response_bytes(mut self, value: u64) -> Self {
-        self.set_max_response_bytes(value);
-        self
-    }
-}
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 struct MaxResponseBytesExtension(pub u64);
@@ -53,25 +35,6 @@ impl MaxResponseBytesRequestExtension for http::request::Builder {
     fn get_max_response_bytes(&self) -> Option<u64> {
         self.extensions_ref()
             .and_then(|extensions| extensions.get::<MaxResponseBytesExtension>().map(|e| e.0))
-    }
-}
-
-/// Add support for transform context to specify how the response will be canonicalized by the replica
-/// to maximize chances of consensus.
-///
-/// See the [docs](https://internetcomputer.org/docs/references/https-outcalls-how-it-works#transformation-function)
-/// on HTTPs outcalls for more details.
-pub trait TransformContextRequestExtension: Sized {
-    /// Set the transform context.
-    fn set_transform_context(&mut self, value: TransformContext);
-
-    /// Retrieve the current transform context, if any.
-    fn get_transform_context(&self) -> Option<&TransformContext>;
-
-    /// Convenience method to use the builder pattern.
-    fn transform_context(mut self, value: TransformContext) -> Self {
-        self.set_transform_context(value);
-        self
     }
 }
